LayoutTests/http/tests/security/cross-frame-access-get-custom-property-cached-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-access-get-custom-property-cached-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-get-custom-property-cached.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-access-get-custom-property-cached-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-get-custom-property-cached.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-access-get-custom-property-cached-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-get-custom-property-cached.html. Domains, protocols and ports must match.

This test checks that caching of property accesses doesn't allow for illegal cross-frame access of a custom property. It also checks for an assertion failure that once occurred in this situation.


PASS: getCustomProperty(targetWindow) should be 'undefined' and is.
PASS: getCustomProperty(targetWindow) should be 'undefined' and is.
PASS: getCustomProperty(targetWindow) should be 'undefined' and is.
